Fix 32ビット考慮

+++ src/app/gkill/dao/account_state/latest_data_repository_address_dao_sqlite3_impl.go
@@ -313,7 +313,7 @@ WHERE TARGET_ID = ?
 	return latestDataRepositoryAddresses[0], nil
 }
 
-func (l *latestDataRepositoryAddressSQLite3Impl) GetLatestDataRepositoryAddressByUpdateTimeAfter(ctx context.Context, updateTime time.Time, limit int) (map[string]*LatestDataRepositoryAddress, error) {
+func (l *latestDataRepositoryAddressSQLite3Impl) GetLatestDataRepositoryAddressByUpdateTimeAfter(ctx context.Context, updateTime time.Time, limit int64) (map[string]*LatestDataRepositoryAddress, error) {
 	// l.m.Lock()
 	// defer l.m.Unlock()
 	sql := fmt.Sprintf(`
